requirements.txt
fhir.resources>=6.2.2
FHIRGenerator>=0.0.10
orjson>=3.6.8
pytest>=7.1.1